+++ app/javascript/vue/tasks/nomenclature/filter/components/filter.vue
@@ -66,6 +66,10 @@
         class="margin-medium-bottom"
         v-model="params.base.taxon_name_relationship_type"
         :nomenclature-code="params.base.nomenclature_code"/>
+      <FacetMatchIdentifiers
+        class="margin-medium-bottom"
+        v-model="params.matchIdentifiers"
+      />
       <tags-component
         class="margin-medium-bottom"
         target="TaxonName"
@@ -116,6 +120,8 @@ import TagsComponent from 'tasks/sources/filter/components/filters/tags'
 import WithComponent from 'tasks/sources/filter/components/filters/with'
 import AuthorsComponent from './filters/authors.vue'
 import FacetDataAttribute from 'tasks/collection_objects/filter/components/filters/DataAttributes/FacetDataAttribute.vue'
+import FacetMatchIdentifiers from 'tasks/people/filter/components/Facet/FacetMatchIdentifiers.vue'
+import checkMatchIdentifiersParams from 'tasks/people/filter/helpers/checkMatchIdentifiersParams'
 
 import SpinnerComponent from 'components/spinner'
 import platformKey from 'helpers/getPlatformKey.js'
@@ -142,7 +148,8 @@ export default {
     UsersComponent,
     TagsComponent,
     WithComponent,
-    FacetDataAttribute
+    FacetDataAttribute,
+    FacetMatchIdentifiers
   },
 
   emits: [
@@ -161,7 +168,7 @@ export default {
       return keys
     },
     parseParams () {
-      const params = Object.assign({}, this.filterEmptyParams(this.params.taxon), this.params.dataAttributes, this.params.authors, this.params.with, this.params.keywords, this.params.related, this.params.base, this.params.user, this.params.includes, this.params.settings)
+      const params = Object.assign({}, checkMatchIdentifiersParams(this.params.matchIdentifiers), this.filterEmptyParams(this.params.taxon), this.params.dataAttributes, this.params.authors, this.params.with, this.params.keywords, this.params.related, this.params.base, this.params.user, this.params.includes, this.params.settings)
       params.updated_since = params.updated_since ? this.setDays(params.updated_since) : undefined
       return params
     }
@@ -251,6 +258,11 @@ export default {
           taxon_name_relationship: [],
           taxon_name_relationship_type: [],
           taxon_name_classification: []
+        },
+        matchIdentifiers: {
+          match_identifiers: undefined,
+          match_identifiers_delimiter: ',',
+          match_identifiers_type: 'internal'
         },
         dataAttributes: {
           data_attribute_value: [],